+++ index.js
@@ -20,7 +20,7 @@ let _opt = {
     digits: 4,
     couples: 4,
     randomUpper: false,
-    numberPosition: 'end' // start|middle|end
+    numberPosition: 'end' // start|middle|end|random
 };
 
 /**
@@ -100,14 +100,18 @@ function generate(opt) {
         str += consonant + vowel;
     }
 
-    if(_opt.digits > 0) {
-        if(_opt.numberPosition === 'start')
+    if (_opt.digits > 0) {
+        if (_opt.numberPosition === 'start')
             str = randomNumber(_opt.digits) + str;
-        else if(_opt.numberPosition === 'middle') {
+        else if (_opt.numberPosition === 'middle') {
             let startPos = Math.round(str.length / 2);
             let number = randomNumber(_opt.digits).toString();
             str = [str.slice(0, startPos), number, str.slice(startPos)].join('');
-        }else if(_opt.numberPosition === 'end')
+        } else if (_opt.numberPosition === 'random') {
+            let startPos = Math.floor(Math.random() * (str.length + 1));
+            let number = randomNumber(_opt.digits).toString();
+            str = [str.slice(0, startPos), number, str.slice(startPos)].join('');
+        } else if (_opt.numberPosition === 'end')
             str += randomNumber(_opt.digits);
         else
             throw new Error('numberPosition can be equal to "start" or "end"');